🛡️ Sentinel: [security improvement] Harden unique ID generation
This PR centralizes and hardens the generation of unique identifiers across the application.

Key changes:
- Introduced a cryptographically secure `generateSecureId` utility in `src/lib/crypto-utils.ts` that uses `crypto.randomUUID()` with secure fallbacks.
- Refactored multiple modules (session analytics, A/B testing, event bus, etc.) to use the new secure utility instead of `Math.random()`.
- Replaced an insecure `console.log` in `src/lib/session-analytics.ts` with a proper logger call to prevent data leakage in non-production environments.
- Improved identifier consistency and entropy for sessions, correlation IDs, and subscriptions.
- Isolated crypto utilities to prevent circular dependencies in the config layer.

Verified with `pnpm security:check`, circular dependency check, and a successful Cloudflare Workers build.
